cleanup

+++ cute_kernels/kernels/continuous_count/cuda_implementation/forward.cu
@@ -43,7 +43,7 @@ inline __device__ void _initialize_global_output(uint32 *output, const uint32 &C
 template <typename scalar_t>
 inline __device__ void _update_local_count(const scalar_t *x,
                                            uint32 *shared_memory,
-                                           const uint32 &num_elements,
+                                           const uint64 &num_elements,
                                            const uint32 &global_thread_id) {
     const uint32 num_elements_per_thread = 16 / sizeof(scalar_t);
     const uint32 num_elements4 = num_elements / num_elements_per_thread;
@@ -70,7 +70,7 @@ inline __device__ void _update_local_count(const scalar_t *x,
 
 template <typename scalar_t>
 __global__ void _continuous_count_cuda_kernel(
-    const scalar_t *x, uint32 *output, const uint32 num_elements, const uint32 C, const bool initialize_output) {
+    const scalar_t *x, uint32 *output, const uint64 num_elements, const uint32 C, const bool initialize_output) {
     const uint32 local_thread_id = ck::get_local_thread_id();
     const uint32 global_thread_id = ck::get_global_thread_id();
     const uint32 num_loops_C = ck::ceil_divide<uint32>(C, blockDim.x);
